cleanup and move proper functionality to proper locations
