+++ src/main/java/com/prosilion/scdecisionmatrix/security/entity/AuthUserDetails.java
@@ -1,4 +1,4 @@
-package com.prosilion.scdecisionmatrix.security;
+package com.prosilion.scdecisionmatrix.security.entity;
 
 import java.util.Collection;
 import org.springframework.security.core.GrantedAuthority;
