+++ src/main/java/com/prosilion/scdecisionmatrix/security/service/AuthUserDetailsService.java
@@ -1,5 +1,6 @@
-package com.prosilion.scdecisionmatrix.security;
+package com.prosilion.scdecisionmatrix.security.service;
 
+import com.prosilion.scdecisionmatrix.security.entity.AuthUserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
 public interface AuthUserDetailsService extends UserDetailsService {
