+++ src/test/resources/application.properties
@@ -12,8 +12,8 @@ spring.jpa.hibernate.ddl-auto=create-drop
 
 # TODO: below needs proper configuration, not working as expected
 #   for now, user creation done in WebSecurityConfig
-spring.security.user.name=superuser
-spring.security.user.password=superpass
+spring.security.user.name=configfileuser
+spring.security.user.password=configfilepass
 spring.security.user.roles=USER
 
 #spring.sql.init.mode=always
